requirements.txt
asgiref==3.3.1
Django==2.2
django-colorful==1.3
django-debug-toolbar==3.2
django-environ==0.4.5
django-mathfilters==1.0.0
django-taggit==1.3.0
djangorestframework==3.12.2
environ==1.0
flake8==3.8.4
gunicorn==20.0.4
mccabe==0.6.1
Pillow==8.0.1
pycodestyle==2.6.0
pyflakes==2.2.0
python-env==1.0.0
pytz==2020.5
sorl-thumbnail==12.7.0
sqlparse==0.4.1
